requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2
pandas==1.5.3
numpy==1.23.5
pandasai
openai>=1.0.0
anthropic>=0.7.0
httpx
streamlit==1.22.0
requests
paramiko
sshtunnel
matplotlib
sentence-transformers
faiss-cpu
xlsxwriter
python-dotenv